image save

+++ loop3.py
@@ -292,8 +292,8 @@ def loop3(config, writer, hydra_log_dir):
         
         for batch in tqdm(train_iter):
             x,y = batch
-            x = x.to(config.device)
-            y = y.to(config.device)
+            x = x.to(config.device, dtype=torch.float32)
+            y = y.to(config.device, dtype=torch.float32)
             
             optimizer.zero_grad()
             #with autocast():
